search-mcp-server/requirements.txt
# MCP Server Requirements for EPIC Search Tools

# Core MCP library
mcp>=1.0.0

# HTTP client for API calls
httpx>=0.25.0

# Web framework for health endpoints
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment variables
python-dotenv>=1.0.0

# Async support
asyncio-mqtt>=0.11.0

# Logging and utilities
structlog>=23.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
